requirements.txt
google-generativeai 
SpeechRecognition
pyttsx3
langchain_community
langchain_google_genai
python-dotenv
langchain
duckduckgo-search
PyAudio
setuptools
psutil
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client